Change package for source code highlighting used
Instead of lstlisting package use the minted package with python
pygments. This requires an additional option "-shell-escape" for
pdflatex which increases danger, however it is worth it, since the
pygments package allows for source code highlighting of many more
languages. In accordance with slightly different package options and
settings relevant changes were made.